+++ admin/src/components/user/edituser.js
@@ -1,9 +1,11 @@
 import React,{useState,useEffect,useContext} from 'react'
 import {usercontext} from '../../App'
 import {serverurl} from '../../config'
-import {Link, useParams} from 'react-router-dom'
+import {Link, useParams,useHistory} from 'react-router-dom'
 import Loading from '../loading'
 import Usernav from '../usernav'
+import confirm from "reactstrap-confirm";
+
 import { scryRenderedDOMComponentsWithClass } from 'react-dom/test-utils'
 const UserEdit= ()=>{
 
@@ -19,9 +21,8 @@ const UserEdit= ()=>{
    const [shop,setshop]=useState(false)
    const [loading,setloading]=useState(false)
     const [pass,setpass]=useState("")
-
+    const history=useHistory()
     useEffect(()=>{
-        console.log(username)
         fetch(`${serverurl}/users/${username}`,{
             method:"get",
             headers:{
@@ -29,14 +30,15 @@ const UserEdit= ()=>{
             }
          }).then(res=>res.json())
          .then(result=>{
-             setdata(result)
-             console.log(result)
+            if(result.err) setmsg("error loading")
+             else{setdata(result)
             setname(result.name)
             setemail(result.email)
             setmobile(result.mobileNumber)
             setlivein(result.livingIn)
             setshop(result.shopping)
             setslot(result.slotbooking)
+             }
             setloading(true)
          }).catch(err=>{
              setloading(true)
@@ -66,8 +68,19 @@ const toggleorder=()=>{
     }
 }
 
-const edituser=()=>{
-    setloading(false)
+const edituser=async ()=>{
+    setmsg("")
+    let result = await confirm(
+        {
+            title: ( "dear admin"),
+            message: "do you really want to change this user details?",
+            confirmText: "ok",
+            confirmColor: "primary",
+            cancelText: "cancel"
+        }
+    ); 
+    if(result==false) return false
+    else setloading(false)
     fetch(`${serverurl}/users/${username}`,{
         method:"put",
         headers:{
@@ -79,18 +92,38 @@ const edituser=()=>{
            liveIn:live,slotbooking:slot,shopping:shop
         })
      }).then(res=>res.json())
-     .then(result=>{
-        console.log(result)
+     .then(async result=>{
+        if(result.err) setmsg("error updating user")
+        else {setmsg("user updated successfully")
+        let result = await confirm(
+            {
+                title: ( "dear admin"),
+                message: "user details are updated",
+                confirmText: "ok",
+                confirmColor: "primary",
+                cancelText: ""
+            }
+        ); }
         setloading(true)
-        setmsg("user updated successfully")
      }).catch(err=>{
-        console.log(err)
         setloading(true)
         setmsg("error updating user")
      })
 }
 
-const deluser=()=>{
+const deluser=async ()=>{
+    setmsg("")
+    let result = await confirm(
+        {
+            title: ( "dear admin"),
+            message: "do you really want to delete this user?",
+            confirmText: "ok",
+            confirmColor: "primary",
+            cancelText: "cancel"
+        }
+    ); 
+    if(result==false) return false
+    else setloading(false)
     fetch(`${serverurl}/users/${username}`,{
         method:"delete",
         headers:{
@@ -98,19 +131,44 @@ const deluser=()=>{
            "Authorization":"Bearer "+localStorage.getItem("token")
         }
      }).then(res=>res.json())
-     .then(result=>{
-        console.log(result)
-        setmsg("user deleted successfully")
+     .then(async (result)=>{
+        if(result.err){setmsg("error deleting user")
+        setloading(true)
+        }else{
+            let result = await confirm(
+                {
+                    title: ( "dear admin"),
+                    message: "the user is deleted",
+                    confirmText: "ok",
+                    confirmColor: "primary",
+                    cancelText: ""
+                }
+            ); 
+            
+            history.push('/userlist')
+        }
      }).catch(err=>{
-        console.log(err)
+        setloading(true)
         setmsg("error deleting user")
      })
 }
-const resetpass=()=>{
+const resetpass=async()=>{
+    setmsg("")
     if(pass.length<8){
         setmsg("password length should be 8")
         return false;
     }
+    let result = await confirm(
+        {
+            title: ( "dear admin"),
+            message: "do you really want to reset password of this user?",
+            confirmText: "ok",
+            confirmColor: "primary",
+            cancelText: "cancel"
+        }
+    ); 
+    if(result==false) return false
+    else setloading(false)
     fetch(`${serverurl}/users/resetpwd`,{
         method:"post",
         headers:{
@@ -121,11 +179,22 @@ const resetpass=()=>{
            username:username,pass:pass
         })
      }).then(res=>res.json())
-     .then(result=>{
-        console.log(result)
-        setmsg("password reset successfully")
+     .then(async result=>{
+        if(result.err) setmsg("error resetting password")
+        else {setpass("")
+        let result = await confirm(
+            {
+                title: ( "dear admin"),
+                message: "password is resetted successfully",
+                confirmText: "ok",
+                confirmColor: "primary",
+                cancelText: ""
+            }
+        );
+        setmsg("password reset successfully") }
+        setloading(true)
      }).catch(err=>{
-        console.log(err)
+        setloading(true)
         setmsg("error resetting password")
      })
 }
@@ -149,7 +218,7 @@ return(
     <button onClick={()=>resetpass()}>reset password</button>
     </div>
     <div><button onClick={()=>edituser()}>submit</button></div>
-    <div><button className='redbutton' onClick={()=>{if(window.confirm('are you sure, you want to delete this user?')){deluser()}}}>delete user</button></div>
+    <div><button className='redbutton' onClick={()=>{deluser()}}>delete user</button></div>
         </div>:
         <Loading />
     }
